+++ src/common/hooks/useResetRecordStatus.ts
@@ -4,7 +4,7 @@ import { RecordStatus } from '@common/constants/record.constants';
 import { useStatusState } from '@src/store';
 
 export const useResetRecordStatus = () => {
-  const { recordStatus, setRecordStatus } = useStatusState();
+  const { recordStatus, setRecordStatus } = useStatusState(['recordStatus', 'setRecordStatus']);
   const [prevResourceId, setPrevResourceId] = useState<string | null>(null);
   const { resourceId } = useParams();
   const setRecordStatusAsOpen = () => setRecordStatus({ type: RecordStatus.open });